Finalizacao Motor

+++ gpio.h
@@ -4,13 +4,13 @@
 #include <stdint.h>
 
 void GPIO_Init(void);
-void PortK_Output(uint8_t valor);
-void PortM_Output_LCD(uint8_t valor);
-void PortM_Output_Keyboard(uint8_t value);
+void PortK_Output(uint32_t valor);
+void PortM_Output_LCD(uint32_t valor);
+void PortM_Output_Keyboard(uint32_t value);
 uint32_t Return_PortM(void);
 uint32_t PortL_Input(void);
-void SetOneExitM(uint8_t offset);
+void SetOneExitM(uint32_t offset);
 uint32_t Return_PortL(void);
-void PortH_Output(uint8_t valor);
+void PortH_Output(uint32_t valor);
 
 #endif
